[@idlebox/native-executer] 添加自我构建和部署功能，防止CI出现异步冲突 [@idlebox/node] 优化数据流处理，简化流的 promise 处理逻辑 [@idlebox/named-pipe] 实现命名管道的读写功能
更新 pnpm 配置，调整依赖项的链接方式，优化工作区的包管理

+++ @idlebox/named-pipe/src/posix.ts
@@ -1,14 +1,24 @@
-import { isNotExistsError } from '@idlebox/common';
-import { execa } from 'execa';
+import { isNotExistsError, timeout } from '@idlebox/common';
+import { execa, type Options } from 'execa';
 import { constants } from 'node:fs';
 import { stat, unlink } from 'node:fs/promises';
 import { resolve } from 'node:path';
 import { PassThrough, type Readable, type Writable } from 'node:stream';
 import { NamedPipeBase, type IOptions } from './fn/abs.js';
-import { FileDescriptor } from './fn/always-open.js';
-import { BufferStream } from './fn/buffer-stream.js';
+import { FileDescriptor } from './fn/posix/file-descriptor.js';
+import { ClosedBy } from './fn/type.js';
 
-export class NamedPipePosix extends NamedPipeBase {
+function normalizeMode(mode: number | undefined): string {
+	if (mode === undefined) {
+		mode = 0o666;
+	} else {
+		mode = mode & 0o777;
+	}
+	// biome-ignore lint/style/useTemplate: simple
+	return '0' + mode.toString(8);
+}
+
+abstract class NamedPipePosixCreate extends NamedPipeBase {
 	public readonly mode;
 
 	constructor(path: string, options: IOptions = {}) {
@@ -18,17 +28,9 @@ export class NamedPipePosix extends NamedPipeBase {
 
 		super(resolve(path), options);
 
-		let mode = options.mode;
-		if (mode === undefined) {
-			mode = 0o666;
-		} else {
-			mode = mode & 0o777;
-		}
-		// biome-ignore lint/style/useTemplate: simple
-		this.mode = '0' + mode.toString(8);
+		this.mode = normalizeMode(options.mode);
 	}
 
-	private isCreateByMe = false;
 	protected override async _create() {
 		try {
 			const ss = await stat(this.path);
@@ -36,11 +38,11 @@ export class NamedPipePosix extends NamedPipeBase {
 				this.logger.verbose`文件已存在且是FIFO`;
 				return;
 			} else {
-				if (ss.size > 0) {
-					throw new Error(`文件存在但不是FIFO: ${this.path}`);
-				} else {
-					this.logger.verbose`删除空的非FIFO文件`;
+				if (ss.isFile() && ss.size === 0) {
+					this.logger.verbose`删除空文件`;
 					await unlink(this.path);
+				} else {
+					throw new Error(`文件存在但不是FIFO: ${this.path}`);
 				}
 			}
 		} catch (e: unknown) {
@@ -51,9 +53,15 @@ export class NamedPipePosix extends NamedPipeBase {
 
 		this.logger.verbose`mkfifo ${this.path} --mode ${this.mode}`;
 		await execa`mkfifo ${this.path} --mode ${this.mode}`;
-		this.isCreateByMe = true;
-	}
 
+		this.registerInner(async () => {
+			this.logger.verbose`删除创建的管道文件: ${this.path}`;
+			await unlink(this.path).catch(ignoreNotExistsError);
+		});
+	}
+}
+
+export class NamedPipePosixPure extends NamedPipePosixCreate {
 	protected override async _read(): Promise<Readable> {
 		// 打开文件读取，模仿 tail -f 的行为，EOF时立刻重新打开
 		const passthrough = new PassThrough({ objectMode: true, highWaterMark: 0 });
@@ -93,20 +101,18 @@ export class NamedPipePosix extends NamedPipeBase {
 		const passthrough = new PassThrough({ objectMode: true, highWaterMark: 0 });
 		const { promise, reject, resolve } = Promise.withResolvers<void>();
 
-		const fdHolder = new FileDescriptor(this.path, constants.O_WRONLY | constants.O_NONBLOCK, this.logger);
+		const fdHolder = new FileDescriptor(this.path, constants.O_WRONLY | constants.O_NONBLOCK, this.logger, false);
 		this.registerInner(fdHolder);
-
-		const bufferStream = passthrough.pipe(new BufferStream(), { end: true });
 
 		fdHolder.onOpen((socket) => {
 			this.logger.verbose`文件已打开，启动管道`;
-			bufferStream.pipe(socket);
+			passthrough.pipe(socket);
 		});
 		fdHolder.onOpen.once(() => resolve());
 
 		fdHolder.onClose((socket) => {
 			this.logger.verbose`文件已关闭`;
-			bufferStream.unpipe(socket);
+			passthrough.unpipe(socket);
 		});
 		fdHolder.onError((e) => {
 			this.logger.error`打开文件失败: ${e}`;
@@ -124,14 +130,119 @@ export class NamedPipePosix extends NamedPipeBase {
 
 		return passthrough;
 	}
+}
 
-	protected override async _close(_graceful: boolean): Promise<void> {
-		if (this.isCreateByMe) {
-			this.logger.verbose`删除管道文件: ${this.path}`;
-			await unlink(this.path).catch(ignoreNotExistsError);
-		} else {
-			this.logger.verbose`不删除管道文件: ${this.path} (不是我创建的)`;
+/**
+ * 调用tail -f读取
+ * 调用cat写入
+ */
+export class NamedPipePosix extends NamedPipePosixCreate {
+	private execute<T extends Options>(cmd: string[], options: T) {
+		const passthrough = new PassThrough({ objectMode: true, highWaterMark: 0 });
+
+		const p = execa({
+			...options,
+			buffer: false,
+			reject: false,
+			encoding: 'buffer',
+			stderr: this.logger.verbose.isEnabled ? 'pipe' : 'ignore',
+		})`${cmd}`;
+
+		let killSent = false;
+		const doKill = async () => {
+			if (killSent) return;
+			killSent = true;
+			p.kill();
+			try {
+				await Promise.race([p, timeout(1000)]);
+			} catch {
+				this.logger.warn`commandline<${cmd}>未能在1秒内退出`;
+				p.kill(9);
+			}
+		};
+		let expectedExit = false;
+
+		this.registerInner(async () => {
+			expectedExit = true;
+			if (killSent) return;
+
+			this.logger.verbose`控制关闭，终止命令行`;
+			doKill();
+		});
+
+		this.logger.debug`程序commandline<${cmd}>启动，pid=${p.pid}`;
+		if (p.stderr) {
+			p.stderr.on('data', (chunk) => {
+				this.logger.verbose`[stderr] long<${chunk.toString()}>`;
+			});
 		}
+
+		p.then(
+			(result) => {
+				killSent = true;
+				this.triggerClose(ClosedBy.Target);
+
+				this.logger.debug`程序${cmd[0]}退出: expectedExit=${expectedExit}, exitCode=${result.exitCode}`;
+				if (expectedExit || result.exitCode === 0) {
+					passthrough.end();
+				} else {
+					this.logger.verbose`通知Broken pipe`;
+					const e = new Error(`Broken pipe file: ${this.path}`);
+					Object.assign(e, { code: 'EPIPE', path: this.path });
+					passthrough.destroy(e);
+				}
+			},
+			() => {
+				console.error(`execa with reject: false should never reject`);
+				process.exit(1);
+			},
+		);
+
+		return { p, passthrough };
+	}
+
+	protected override async _read(): Promise<Readable> {
+		const { p, passthrough } = this.execute(['tail', '-f', this.path], {
+			stdin: 'ignore',
+			stdout: 'pipe',
+		});
+
+		// 理论上不应该用pipe
+		p.stdout.pipe(passthrough, { end: false });
+
+		passthrough.on('finish', () => {
+			this.logger.debug`写入流关闭`;
+		});
+
+		passthrough.on('end', () => {
+			this.logger.debug`读取流关闭`;
+			this.triggerClose(ClosedBy.Consumer);
+		});
+
+		return passthrough;
+	}
+
+	protected override async _write(): Promise<Writable> {
+		const { p, passthrough } = this.execute(['tee', this.path], {
+			stdin: 'pipe',
+			stdout: 'ignore',
+		});
+
+		p.stdin.on('error', (e) => {
+			this.logger.error`stdin error: ${e}`;
+		});
+		passthrough.pipe(p.stdin, { end: true });
+
+		passthrough.on('finish', () => {
+			this.logger.debug`写入流关闭`;
+			this.triggerClose(ClosedBy.Consumer);
+		});
+
+		passthrough.on('end', () => {
+			this.logger.debug`读取流关闭`;
+		});
+
+		return passthrough;
 	}
 }
